# backend/docs/openapi.yaml
openapi: 3.1.1
info:
  title: Dragons Counter API
  description: 中日ドラゴンズ観戦記録API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.dragons-counter.com
    description: Production server
paths:
...
  /games/{gameId}:
    get:
      tags:
        - Games
      summary: 観戦試合詳細取得
      operationId: getGameById
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: 見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Games
      summary: 観戦試合更新
      operationId: updateGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: 見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Games
      summary: 観戦試合削除
      description: 指定された観戦試合を論理削除します。物理的にデータは削除されず、deletedAtフィールドに削除日時が記録されます。
      operationId: deleteGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 削除成功
        '404':
          description: 見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Game:
      $ref: './schemas/Game.yaml'
...
    UpdateGameRequest:
      $ref: './schemas/UpdateGameRequest.yaml'
...
    Error:
      $ref: './schemas/Error.yaml'